requirements.txt
fastapi
uvicorn
motor
requests
redis
pydantic-settings
jinja2
psutil
pyyaml